requirements.txt
google-cloud-aiplatform>=1.1.1
google-cloud-bigquery>=2.20.0
pandas>=1.2.5
numpy>=1.21.0
pyarrow>=5.0.0
scipy>=1.7.0
kfp>=1.8.9
google-cloud-pipeline-components>=0.1.4
pillow>=10.0.1
matplotlib>=3.4.3
scipy>=1.7.1
seaborn>=0.11.2
scikit-learn>=0.24.2
tensorflow~=2.7
db-dtypes>=1.0.0
ipykernel>=6.15.0